Rename platform identifiers for each Flutter app
