+++ apps/admin_app/pubspec.yaml
@@ -1,4 +1,4 @@
-name: pilot_app
+name: admin_app
 description: A new Flutter project.
 publish_to: 'none'
 version: 1.0.0+1
