+++ apps/cleaner_app/android/app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
     <application
-        android:label="pilot_app"
+        android:label="Cleaner App"
         android:name="${applicationName}"
         android:icon="@mipmap/ic_launcher">
         <activity
