+++ apps/cleaner_app/android/app/src/main/kotlin/com/yourco/cleaner/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.pilot.pilot_app
+package com.yourco.cleaner
 
 import io.flutter.embedding.android.FlutterActivity
