+++ apps/guest_app/android/app/src/main/kotlin/com/yourco/guest/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.pilot.pilot_app
+package com.yourco.guest
 
 import io.flutter.embedding.android.FlutterActivity
